refactoring and fixed some exception errors
