+++ lib/app/explore/start_page.dart
@@ -22,7 +22,7 @@ import 'package:provider/provider.dart';
 import 'package:shimmer/shimmer.dart';
 import 'package:snapd/snapd.dart';
 import 'package:software/snapx.dart';
-import 'package:software/app/installed/installed_snaps_page.dart';
+import 'package:software/app/common/error_page.dart';
 import 'package:software/app/common/loading_banner_grid.dart';
 import 'package:software/app/common/snap/snap_section.dart';
 import 'package:software/app/explore/explore_model.dart';
@@ -30,6 +30,7 @@ import 'package:software/app/explore/section_banner.dart';
 import 'package:software/app/explore/section_grid.dart';
 import 'package:software/services/snap_service.dart';
 import 'package:yaru_colors/yaru_colors.dart';
+import 'package:yaru_icons/yaru_icons.dart';
 
 class StartPage extends StatefulWidget {
   const StartPage({
@@ -86,7 +87,9 @@ class _StartPageState extends State<StartPage> {
     }
 
     if (!SnapService.isSnapdInstalled) {
-      return NoSnapsPage(message: "Snapd is not installed on your system");
+      return const ErrorPage(
+          icon: YaruIcons.warning,
+          message: "Snapd is not installed on your system");
     }
 
     if (bannerSnap == null || bannerSnap2 == null || bannerSnap3 == null) {
